front-end compliant edits

+++ controllers/appointment/appointment.ts
@@ -31,9 +31,10 @@ export const addAppointment = async (
 
     if (!moment(date).isValid()) return res.status(400).json("Invalid date");
 
-    const dayDifferece = moment(date).diff(moment(), "days");
+    const dayDifferece = moment(`${date} ${time}`).diff(moment(), "minutes");
 
-    if (dayDifferece < 0) return res.status(400).json("Invalid date");
+    if (dayDifferece < 0)
+      return res.status(400).json("Cannot schedule past dates");
 
     const doesUserExist = await prisma.user.findUnique({
       where: {
@@ -83,6 +84,9 @@ export const addAppointment = async (
       return res.status(400).json("Time must be in increments of 20 minutes");
     }
     const weekDayWorkingHours = JSON.parse((vetCalender as any)[weekDay]);
+
+    if(weekDayWorkingHours === "closed") return res.status(400).json("Vet is not available on this day");
+
 
     const isBetweenWorkingHours =
       moment(`${date}  ${time}`).isBetween(
@@ -168,9 +172,10 @@ export const updateAppointment = async (
 
     if (!moment(date).isValid()) return res.status(400).json("Invalid date");
 
-    const dayDifferece = moment(date).diff(moment(), "days");
+    const dayDifferece = moment(`${date} ${time}`).diff(moment(), "minutes");
 
-    if (dayDifferece < 0) return res.status(400).json("Invalid date");
+    if (dayDifferece < 0)
+      return res.status(400).json("Cannot schedule past dates");
 
     const doesUserExist = await prisma.user.findUnique({
       where: {
@@ -232,6 +237,8 @@ export const updateAppointment = async (
     }
     const weekDayWorkingHours = JSON.parse((vetCalender as any)[weekDay]);
 
+    if(weekDayWorkingHours === "closed") return res.status(400).json("Vet is not available on this day");
+    
     const isBetweenWorkingHours =
       moment(`${date}  ${time}`).isBetween(
         `${date}  ${weekDayWorkingHours.morning.start_at}`,
@@ -446,6 +453,13 @@ export const closeTimeSlot = async (
     });
 
     if (existingAppointment) {
+      if (
+        existingAppointment.user_id === process.env.DEFAULT_USER_ID &&
+        existingAppointment.pet_id === process.env.DEFAULT_PET_ID
+      ) {
+        return res.status(400).json("Time slot is already closed");
+      }
+
       const deletedAppointment = await prisma.appointment.delete({
         where: {
           id: existingAppointment.id,
@@ -580,3 +594,342 @@ export const openTimeSlot = async (
     console.log(e);
   }
 };
+
+export const addAppointmentVet = async (
+  req: Request,
+  res: Response,
+  prisma: PrismaClient
+) => {
+  try {
+    const { date, time, pet_id, vet_id, user_id } = req.body;
+
+    if (
+      date == undefined ||
+      time == undefined ||
+      pet_id == undefined ||
+      vet_id == undefined ||
+      user_id == undefined
+    )
+      return res.status(400).json("Missing fields");
+
+    const payload: JWTPayload = handleTokenVerification(req, res) as JWTPayload;
+
+    if (payload.userId != vet_id) return res.status(401).json("Unauthorized");
+
+    if (!moment(date).isValid()) return res.status(400).json("Invalid date");
+
+    const dayDifferece = moment(`${date} ${time}`).diff(moment(), "minutes");
+
+    if (dayDifferece < 0)
+      return res.status(400).json("Cannot schedule past dates");
+
+    const doesUserExist = await prisma.user.findUnique({
+      where: {
+        id: user_id,
+      },
+    });
+
+    if (!doesUserExist)
+      return res.status(404).json(`User with id:${user_id} does not exist`);
+
+    const doesPetExist = await prisma.pet.findUnique({
+      where: {
+        id: pet_id,
+      },
+    });
+
+    if (!doesPetExist)
+      return res.status(404).json(`Pet with id:${pet_id} does not exist`);
+
+    const doesVetExist = await prisma.vet.findUnique({
+      where: {
+        id: vet_id,
+      },
+    });
+
+    if (!doesVetExist)
+      return res.status(404).json(`Vet with id:${vet_id} does not exist`);
+
+    if (!doesVetExist.is_approved)
+      return res.status(400).json("Vet is not approved yet");
+
+    const dateValue = new Date(date);
+    const day = dateValue.getDay() as number;
+
+    const weekDay = (weekDays as any)[day];
+
+    const vetCalender = await prisma.calendar.findUnique({
+      where: {
+        owner_id: vet_id,
+      },
+    });
+
+    if (!JSON.parse((vetCalender as any)[weekDay]))
+      return res.status(400).json("Vet is not available on this day");
+
+    if (parseInt(time.split(":")[1]) % 20 != 0) {
+      return res.status(400).json("Time must be in increments of 20 minutes");
+    }
+    const weekDayWorkingHours = JSON.parse((vetCalender as any)[weekDay]);
+
+    if(weekDayWorkingHours === "closed") return res.status(400).json("Vet is not available on this day");
+
+    const isBetweenWorkingHours =
+      moment(`${date}  ${time}`).isBetween(
+        `${date}  ${weekDayWorkingHours.morning.start_at}`,
+        `${date}  ${weekDayWorkingHours.morning.end_at}`
+      ) ||
+      moment(`${date}  ${time}`).isBetween(
+        `${date}  ${weekDayWorkingHours.afternoon.start_at}`,
+        `${date}  ${weekDayWorkingHours.afternoon.end_at}`
+      );
+
+    if (!isBetweenWorkingHours)
+      return res.status(400).json("Vet is not available at this time");
+
+    const existingAppointment = await prisma.appointment.findFirst({
+      where: {
+        vet_id,
+        date,
+        time,
+      },
+    });
+
+    if (existingAppointment)
+      return res.status(400).json("Appointment at that time already exists");
+
+    const newAppointment = await prisma.appointment.create({
+      data: {
+        date,
+        time,
+        pet_id,
+        vet_id,
+        user_id,
+      },
+    });
+
+    const msg = {
+      to: doesUserExist.email, // Change to your recipient
+      from: "info@aimensahnoun.com", // Change to your verified sender
+      subject: "VetoLib Appointment",
+      html: `<div><strong>Dear ${doesUserExist.first_name} ${doesUserExist.last_name}</strong> <span> Has booked an appointment with Doctor ${doesVetExist.first_name} ${doesVetExist.last_name} on ${date} ${time} for pet: ${doesPetExist.name}</span></div>
+    `,
+    };
+
+    sgMail
+      .send(msg)
+      .then(() => {
+        console.log("Email sent");
+      })
+      .catch((error) => {
+        console.error(error);
+      });
+
+    res.status(201).json(newAppointment);
+  } catch (e) {
+    console.log(e);
+  }
+};
+
+export const updateAppointmentVet = async (
+  req: Request,
+  res: Response,
+  prisma: PrismaClient
+) => {
+  try {
+    const { id } = req.params;
+
+    if (!id || id === "") return res.status(400).json("Missing id");
+
+    const { date, time, pet_id, vet_id, user_id } = req.body;
+
+    if (
+      date == undefined ||
+      time == undefined ||
+      pet_id == undefined ||
+      vet_id == undefined ||
+      user_id == undefined
+    )
+      return res.status(400).json("Missing fields");
+
+    const payload: JWTPayload = handleTokenVerification(req, res) as JWTPayload;
+
+    if (payload.userId != vet_id) return res.status(401).json("Unauthorized");
+
+    if (!moment(date).isValid()) return res.status(400).json("Invalid date");
+
+    const dayDifferece = moment(`${date} ${time}`).diff(moment(), "minutes");
+
+    if (dayDifferece < 0)
+      return res.status(400).json("Cannot schedule past dates");
+
+    const doesUserExist = await prisma.user.findUnique({
+      where: {
+        id: user_id,
+      },
+    });
+
+    if (!doesUserExist)
+      return res.status(404).json(`User with id:${user_id} does not exist`);
+
+    const doesPetExist = await prisma.pet.findUnique({
+      where: {
+        id: pet_id,
+      },
+    });
+
+    if (!doesPetExist)
+      return res.status(404).json(`Pet with id:${pet_id} does not exist`);
+
+    const doesVetExist = await prisma.vet.findUnique({
+      where: {
+        id: vet_id,
+      },
+    });
+
+    if (!doesVetExist)
+      return res.status(404).json(`Vet with id:${vet_id} does not exist`);
+
+    if (!doesVetExist.is_approved)
+      return res.status(400).json("Vet is not approved yet");
+
+    const oldAppointment = await prisma.appointment.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!oldAppointment) return res.status(404).json("Appointment not found");
+
+    if (!oldAppointment.vet_id == vet_id)
+      return res.status(401).json("Unauthorized");
+
+    const dateValue = new Date(date);
+    const day = dateValue.getDay() as number;
+
+    const weekDay = (weekDays as any)[day];
+
+    const vetCalender = await prisma.calendar.findUnique({
+      where: {
+        owner_id: vet_id,
+      },
+    });
+
+    if (!JSON.parse((vetCalender as any)[weekDay]))
+      return res.status(400).json("Vet is not available on this day");
+
+    if (parseInt(time.split(":")[1]) % 20 != 0) {
+      return res.status(400).json("Time must be in increments of 20 minutes");
+    }
+    const weekDayWorkingHours = JSON.parse((vetCalender as any)[weekDay]);
+
+    if(weekDayWorkingHours === "closed") return res.status(400).json("Vet is not available on this day");
+
+
+    const isBetweenWorkingHours =
+      moment(`${date}  ${time}`).isBetween(
+        `${date}  ${weekDayWorkingHours.morning.start_at}`,
+        `${date}  ${weekDayWorkingHours.morning.end_at}`
+      ) ||
+      moment(`${date}  ${time}`).isBetween(
+        `${date}  ${weekDayWorkingHours.afternoon.start_at}`,
+        `${date}  ${weekDayWorkingHours.afternoon.end_at}`
+      );
+
+    if (!isBetweenWorkingHours)
+      return res.status(400).json("Vet is not available at this time");
+
+    const newAppointment = await prisma.appointment.update({
+      where: {
+        id,
+      },
+      data: {
+        date,
+        time,
+        pet_id,
+        vet_id,
+        user_id,
+      },
+    });
+
+    const msg = {
+      to: doesUserExist.email, // Change to your recipient
+      from: "info@aimensahnoun.com", // Change to your verified sender
+      subject: "VetoLib Appointment Update",
+      html: `<div><strong>Dear ${doesUserExist.first_name} ${doesUserExist.last_name}</strong> <span> Has booked an appointment with Doctor ${doesVetExist.first_name} ${doesVetExist.last_name} on ${date} ${time} for pet: ${doesPetExist.name}</span></div>
+    `,
+    };
+
+    sgMail
+      .send(msg)
+      .then(() => {
+        console.log("Email sent");
+      })
+      .catch((error) => {
+        console.error(error);
+      });
+
+    res.status(200).json(newAppointment);
+  } catch (e) {
+    console.log(e);
+  }
+};
+
+export const cancelAppointmentVet = async (
+  req: Request,
+  res: Response,
+  prisma: PrismaClient
+) => {
+  try {
+    const { id } = req.params;
+
+    if (!id || id === "") return res.status(400).json("Missing id");
+
+    const { vet_id } = req.headers;
+
+    if (vet_id == undefined) return res.status(400).json("Missing user_id");
+
+    const payload: JWTPayload = handleTokenVerification(req, res) as JWTPayload;
+
+    if (payload.userId != vet_id) return res.status(401).json("Unauthorized");
+
+    const doesUserExist = await prisma.vet.findUnique({
+      where: {
+        id: vet_id,
+      },
+    });
+
+    if (!doesUserExist) return res.status(404).json("Vet not found");
+
+    const doesAppointmentExist = await prisma.appointment.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!doesAppointmentExist)
+      return res.status(404).json("Appointment not found");
+
+    const appointment = await prisma.appointment.delete({
+      where: {
+        id,
+      },
+    });
+
+    if (!appointment) return res.status(404).json("Appointment not found");
+
+    const msg = {
+      to: doesUserExist.email, // Change to your recipient
+      from: "info@aimensahnoun.com", // Change to your verified sender
+      subject: "VetoLib Appointment Cancelled",
+      html: `<div><strong>Dear ${doesUserExist.first_name} ${doesUserExist.last_name}</strong> <span> Has canceled an appointment with id ${appointment.id} successfully </span></div>
+    `,
+    };
+
+    res.status(200).json("Appointment canceled successfully");
+  } catch (e) {
+    console.log(e);
+  }
+};
+
+
